backend/requirements.txt
Django==4.0
psycopg2-binary==2.9.2
python-decouple==3.5
gunicorn==20.1.0
django-cors-headers==3.10.1
bs4==0.0.1
requests==2.26.0
nltk==3.6.5
django-extensions==3.1.5
djangorestframework==3.13.1